clean: moved cli tests to cli/tests.go
